refactored opportunities.jsx
